backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openpyxl==3.1.2
pandas==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic==2.7.1
